[FEAT] authorization

+++ src/posts/posts.service.ts
@@ -1,19 +1,34 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from './entities/post.entity';
 import { Repository } from 'typeorm';
+import { User } from 'src/users/entities/user.entity';
 
 @Injectable()
 export class PostsService {
   constructor(
     @InjectRepository(Post)
     private postsRepository: Repository<Post>,
+    @InjectRepository(User)
+    private usersRepository: Repository<User>,
   ) {}
 
-  async create(createPostDto: CreatePostDto) {
-    const post = this.postsRepository.create(createPostDto);
+  async create(createPostDto: CreatePostDto, userId: number) {
+    const user = await this.usersRepository.findOne({ where: { id: userId } });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    const post = this.postsRepository.create({
+      ...createPostDto,
+      user: user,
+    });
     return await this.postsRepository.save(post);
   }
 
@@ -21,20 +36,38 @@ export class PostsService {
     return await this.postsRepository.find();
   }
 
-  async findOne(id: number) {
-    const post = await this.postsRepository.findOne({ where: { id } });
+  async findOne(id: number, userId: number) {
+    const post = await this.postsRepository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
 
     if (!post) {
       throw new NotFoundException('Post not found');
     }
 
+    if (post.user.id !== userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to view this post',
+      );
+    }
+
     return post;
   }
 
-  async update(id: number, updatePostDto: UpdatePostDto) {
-    const post = await this.postsRepository.findOne({ where: { id } });
+  async update(id: number, updatePostDto: UpdatePostDto, userId: number) {
+    const post = await this.postsRepository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
     if (!post) {
       throw new NotFoundException('Post not found');
+    }
+
+    if (post.user.id !== userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this post',
+      );
     }
 
     return await this.postsRepository.save({
@@ -43,11 +76,21 @@ export class PostsService {
     });
   }
 
-  async remove(id: number) {
-    const post = await this.postsRepository.findOne({ where: { id } });
+  async remove(id: number, userId: number) {
+    const post = await this.postsRepository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
     if (!post) {
       throw new NotFoundException('Post not found');
     }
+
+    if (post.user.id !== userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to remove this post',
+      );
+    }
+
     return await this.postsRepository.remove(post);
   }
 }
